+++ Memy/Server/Service/UserService.cs
@@ -1,4 +1,5 @@
-﻿using Memy.Server.Data.File;
+﻿using Memy.Server.Data;
+using Memy.Server.Data.File;
 using Memy.Server.Data.User;
 using Memy.Shared.Helper;
 using Memy.Shared.Model;
@@ -7,14 +8,12 @@ namespace Memy.Server.Service
 {
     internal class UserService
     {
-        private readonly ILogger _logger;
         private readonly IUserData _userData;
 
         private readonly IWebHostEnvironment _webHostEnvironment;
 
-        internal UserService(ILogger logger, IUserData userData, IWebHostEnvironment webHostEnvironment)
+        internal UserService(IUserData userData, IWebHostEnvironment webHostEnvironment)
         {
-            _logger = logger;
             _userData = userData;
             _webHostEnvironment = webHostEnvironment;
         }
@@ -27,9 +26,8 @@ namespace Memy.Server.Service
 
                 return result;
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                _logger.LogError(ex.Message);
                 throw;
             }
         }
@@ -41,53 +39,68 @@ namespace Memy.Server.Service
                 var result = await _userData.GetEmail(token);
                 return result;
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                _logger.LogError(ex.Message);
                 throw;
             }
         }
 
-        internal async Task UpdatePassword(string token, string value)
+        internal async Task UpdatePassword(string? token, string? value)
         {
             try
             {
-                var password = ConvertByteString.ConvertToObject<Password>(value);
-                await _userData.UpdatePassword(token, password.Old, password.New);
+                if (!string.IsNullOrWhiteSpace(token) && !string.IsNullOrWhiteSpace(value))
+                {
+                    var password = ConvertByteString.ConvertToObject<Password>(value);
+                    await _userData.UpdatePassword(token, password.Old, password.New);
+                }
             }
-            catch (Exception ex)
+            catch (Exception )
             {
-                _logger.LogError(ex.Message);
                 throw;
             }
         }
 
 
-        internal async Task<bool> UpdateName(string token, string value)
+        internal async Task<bool> UpdateName(string? token, string? value)
         {
             try
             {
-                var available = await _userData.NameIsAvailable(value);
+                bool available = false;
+                if (!string.IsNullOrWhiteSpace(value))
+                {
+                    available = await _userData.NameIsAvailable(value);
+                }
                 if (available)
                 {
-                    await _userData.UpdateName(token, value);
-                    return true;
+                    if (!string.IsNullOrWhiteSpace(token) && !string.IsNullOrWhiteSpace(value))
+                    {
+                        await _userData.UpdateName(token, value);
+                        return true;
+                    }
+                    return false;
                 }
                 else
                 {
                     return false;
                 }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                _logger.LogError(ex.Message);
                 throw;
             }
         }
 
-        internal async Task UpdateAvatar(string token, string value)
+        internal async Task UpdateAvatar(string? token, string? value)
         {
-            var avatar = Newtonsoft.Json.JsonConvert.DeserializeObject<FileUploadStatus>(value);
+            FileUploadStatus? avatar = null;
+
+            if (!string.IsNullOrWhiteSpace(value))
+            {
+                avatar = Newtonsoft.Json.JsonConvert.DeserializeObject<FileUploadStatus>(value);
+            }
+
+
             try
             {
 
@@ -97,18 +110,27 @@ namespace Memy.Server.Service
                 }
                 var status = await CheckingFile.CorrectData(avatar, _webHostEnvironment);
 
-                await _userData.UpdateAvatar(token, status.ObjName);
+                if (status is not null)
+                {
+                    if (!string.IsNullOrWhiteSpace(token) && !string.IsNullOrWhiteSpace(status.ObjName))
+                    {
+                        await _userData.UpdateAvatar(token, status.ObjName);
+                    }
+                }
             }
-            catch (Exception ex)
+            catch (Exception )
             {
-                CheckingFile.DeleteFile(avatar.ObjName, _webHostEnvironment);
-                _logger.LogError(ex.Message);
+                if (avatar is not null)
+                {
+                    CheckingFile.DeleteFile(avatar.ObjName, _webHostEnvironment);
+                }
                 throw;
             }
         }
 
-        internal async Task<string> UpdateEmail(string token, string value)
+        internal async Task<string> UpdateEmail(string? token, string? value)
         {
+            await Task.Run(() => { Task.Delay(1); });
             throw new NotImplementedException();
         }
     }
